+++ package.json
@@ -9,14 +9,13 @@
     "apca-w3": "^0.1.1",
     "chroma-js": "^2.1.2",
     "comlink": "^4.3.1",
-    "dart-sass": "^1.25.0",
     "lodash": "^4.17.21",
     "lz-string": "^1.4.4",
     "nanostores": "^0.5.8",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
-    "styled-components": "^5.3.0",
+    "react-scripts": "^5.0.1",
+    "styled-components": "^5.3.5",
     "typescript": "^4.0.3",
     "web-vitals": "^0.2.4",
     "workbox-background-sync": "^5.1.3",
@@ -30,7 +29,8 @@
     "workbox-range-requests": "^5.1.3",
     "workbox-routing": "^5.1.3",
     "workbox-strategies": "^5.1.3",
-    "workbox-streams": "^5.1.3"
+    "workbox-streams": "^5.1.3",
+    "workerize-loader": "^2.0.2"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -67,13 +67,12 @@
     "@types/node": "^12.0.0",
     "@types/react": "^16.9.53",
     "@types/react-dom": "^16.9.8",
-    "@types/styled-components": "^5.1.11",
-    "sass": "^1.41.1",
+    "@types/styled-components": "^5.1.26",
+    "sass": "^1.54.4",
     "stylelint": "^13.13.1",
     "stylelint-config-recommended": "^5.0.0",
     "stylelint-config-styled-components": "^0.1.1",
     "stylelint-processor-styled-components": "^1.10.0",
-    "use-local-storage-state": "^10.0.0",
-    "worker-loader": "^3.0.8"
+    "use-local-storage-state": "^10.0.0"
   }
 }
