+++ src/components/App.tsx
@@ -8,7 +8,7 @@ import { ExportField } from './Export'
 import { ColorInfo } from './ColorInfo'
 import { Help } from './Help'
 import { KeyPressHandler } from './KeyPressHandler'
-import { useKeyPress } from '../hooks/useKeyPress'
+import { useKeyPress } from '../shared/hooks/useKeyPress'
 import { paletteStore } from '../store/palette'
 import { selectedStore, setSelected } from '../store/currentPosition'
 import { Header } from './Header'
