+++ src/components/ColorGraph/Chart/Canvas.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useMemo, useRef } from 'react'
-import { Channel, TColor } from '../../../types'
+import { Channel, TColor } from '../../../shared/types'
 import debounce from 'lodash/debounce'
 import styled from 'styled-components'
-import { TSpaceName } from '../../../colorFuncs'
+import { TSpaceName } from '../../../shared/colorFuncs'
 import { useStore } from '@nanostores/react'
 import { paletteStore } from '../../../store/palette'
 import { chartSettingsStore } from '../../../store/chartSettings'
