+++ src/components/ColorGraph/Chart/RenderStrategy/WorkerPool/CanvasWorkerPool.ts
@@ -1,9 +1,9 @@
 import { DrawChartProps, WorkerObj } from '../../paintWorker'
-import { Channel } from '../../../../../types'
+import { Channel } from '../../../../../shared/types'
 
 // @ts-ignore Module not found
 // eslint-disable-next-line import/no-webpack-loader-syntax
-import PaintWorker from 'worker-loader!../../paintWorker'
+import PaintWorker from 'workerize-loader!../../paintWorker'
 import * as Comlink from 'comlink'
 
 export type PaintResult = ImageBitmap | ImageData
@@ -57,7 +57,9 @@ class CanvasWorkerPool {
   constructor(size: number = CanvasWorkerPool.optimalPoolSize) {
     const workers = new Array(size).fill(0).map(() => new PaintWorker())
     const comlinks = workers.map(w => Comlink.wrap<WorkerObj>(w))
+
     const funcs = comlinks.map(convertWorkerToFuncs)
+    console.log({ funcs })
 
     this.workers = workers
     this.comlinks = comlinks
