+++ src/components/Header/ThemeButton.tsx
@@ -1,4 +1,4 @@
-import { useColorScheme } from '../../hooks/useColorScheme'
+import { useColorScheme } from '../../shared/hooks/useColorScheme'
 import { Button } from '../inputs'
 import { MoonIcon } from '../../icons/Moon'
 import { SunIcon } from '../../icons/Sun'
