+++ src/store/palette/stores.ts
@@ -1,10 +1,10 @@
-import { Palette } from '../../types'
+import { Palette } from '../../shared/types'
 import { computed, map, onMount } from 'nanostores'
-import { colorSpaces } from '../../colorFuncs'
+import { colorSpaces } from '../../shared/colorFuncs'
 import { parseHexPalette } from '.'
 import { persistentAtom } from '@nanostores/persistent'
 import { PRESETS } from '../../presets'
-import { HexPalette } from '../../types'
+import { HexPalette } from '../../shared/types'
 import { PALETTE_KEY } from '../../constants'
 import { atom } from 'nanostores'
 import { isEqual } from 'lodash'
